requirements-dev.txt
pytest-cov>=5.0.0
requests>=2.31.0
